pyproject.toml
[project]
name = "music_bg"
version = "0.2.1"
description = "Dynamic wallpapers with power of mpris2."
authors = [{ name = "Pavel Kirilin", email = "s3riussan@gmail.com" }]
requires-python = "~=3.13"
readme = "README.md"
keywords = ["Music", "Dbus", "Image Processing", "Music Background", "Mpris2"]
classifiers = [
	"Operating System :: POSIX :: Linux",
	"Topic :: Scientific/Engineering :: Image Processing",
	"Topic :: Multimedia :: Sound/Audio",
	"Typing :: Typed",
]
dependencies = [
	"dbus-python>=1.2.16,<2",
	"PyGObject>=3.40.1,<4",
	"loguru>=0.5.3,<0.6",
	"requests>=2.26.0,<3",
	"pydantic>=1.8.2,<2",
	"Pillow>=12.0.0,<13",
	"toml>=0.10.2",
	"numpy>=2.3.4",
]

[project.urls]
Homepage = "https://github.com/music-bg/music_bg/wiki"
Repository = "https://github.com/music-bg/music_bg/"
Documentation = "https://github.com/music-bg/music_bg/wiki"

[project.scripts]
music_bg = "music_bg.__main__:main"

[project.entry-points.mbg_processors]
fit = "music_bg.img_processors.fit:fit"
noop = "music_bg.img_processors.noop:noop"
resize = "music_bg.img_processors.resize:resize"
load_img = "music_bg.img_processors.load_img:load_img"
blank_image = "music_bg.img_processors.blank_img:blank_image"
box_blur = "music_bg.img_processors.blur:box_blur"
gaussian_blur = "music_bg.img_processors.blur:gaussian_blur"
circle = "music_bg.img_processors.circle:circle"
pop_filter = "music_bg.img_processors.pop:pop_filter"
print = "music_bg.img_processors.print:img_print"
radial_gradient = "music_bg.img_processors.gradients:radial_gradient"

[project.entry-points.mbg_variables]
uuid4 = "music_bg.img_variables.uuid_gen:uuid4"
most_frequent_color = "music_bg.img_variables.colors:most_frequent_color_var"
most_frequent_color_inverted = "music_bg.img_variables.colors:most_frequent_color_inverted_var"
least_frequent_color = "music_bg.img_variables.colors:most_frequent_color_inverted_var"

[dependency-groups]
dev = [
	"mypy>=1,<2",
	"pytest>=8,<9",
	"pre-commit>=4,<5",
	"ruff>=0.14.3,<1",
	"types-requests>=2.32.4.20250913",
	"types-toml>=0.10.8.20240310",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true


[tool.pytest.ini_options]
log_level = 'INFO'
anyio_mode = "auto"

[tool.ruff]
target-version = "py39"
# List of enabled rulsets.
# See https://docs.astral.sh/ruff/rules/ for more information.
lint.select = [
	"E",     # Error
	"F",     # Pyflakes
	"W",     # Pycodestyle
	"C90",   # McCabe complexity
	"I",     # Isort
	"N",     # pep8-naming
	"D",     # Pydocstyle
	"ANN",   # Pytype annotations
	"S",     # Bandit
	"B",     # Bugbear
	"COM",   # Commas
	"C4",    # Comprehensions
	"ISC",   # Implicit string concat
	"PIE",   # Unnecessary code
	"T20",   # Catch prints
	"PYI",   # validate pyi files
	"Q",     # Checks for quotes
	"RSE",   # Checks raise statements
	"RET",   # Checks return statements
	"SLF",   # Self checks
	"SIM",   # Simplificator
	"PTH",   # Pathlib checks
	"ERA",   # Checks for commented out code
	"PL",    # PyLint checks
	"RUF",   # Specific to Ruff checks
	"FA102", # Future annotations
]
lint.ignore = [
	"D105",    # Missing docstring in magic method
	"D107",    # Missing docstring in __init__
	"D212",    # Multi-line docstring summary should start at the first line
	"D401",    # First line should be in imperative mood
	"D104",    # Missing docstring in public package
	"D100",    # Missing docstring in public module
	"ANN401",  # typing.Any are disallowed in `**kwargs
	"PLR0913", # Too many arguments for function call
	"D106",    # Missing docstring in public nested class
]
lint.mccabe = { max-complexity = 10 }
exclude = [".venv/"]
line-length = 88

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
	"S101",   # Use of assert detected
	"S301",   # Use of pickle detected
	"D103",   # Missing docstring in public function
	"SLF001", # Private member accessed
	"S311",   # Standard pseudo-random generators are not suitable for security/cryptographic purposes
	"D101",   # Missing docstring in public class
	"D102",   # Missing docstring in public method
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
ignore-decorators = ["typing.overload"]

[tool.ruff.lint.pylint]
allow-magic-value-types = ["int", "str", "float"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["taskiq_dependencies.Depends", "taskiq.TaskiqDepends"]

[tool.typos.files]
extend-exclude = [
	"docs/README.md", # because of identifier in head section
]
